Applying cloning for loading tables

+++ plugins/plugins/OracleSubject.java
@@ -14,6 +14,7 @@ package plugins;
 
 import oracle.jdbc.OracleResultSet;
 import oracle.sql.CLOB;
+
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.ByteArrayInputStream;
@@ -27,6 +28,8 @@ import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.Vector;
+
+import azdblab.Constants;
 import azdblab.exception.dbms.DBMSInvalidConnectionParameterException;
 import azdblab.executable.Main;
 import azdblab.labShelf.GeneralDBMS;
@@ -668,7 +671,8 @@ public class OracleSubject extends LabShelfManager {
 			for (int i = 0; i < vecTables.size(); i++) {
 				String tblName = (String) vecTables.get(i);
 				Main._logger.outputLog("installed tableName: " + tblName);
-				dropTable(tblName);
+				if(!tblName.contains(Constants.CLONE_TABLE_PREFIX))
+					dropTable(tblName);
 			}
 		} catch (SQLException e) {
 			e.printStackTrace();
